requirements.txt
python-dotenv==0.20.0
ruuvitag-sensor==2.0.0
influxdb==5.3.1
